Mega jumbo tent update
Added host transfer command
Added automatic host transfers
Added logging for users joining and leaving for ease of reporting and such
Added /add to whitelist users so users can invite others to a locked tent without having to unlock it
Added /limit to give users better control over an open tent, as overcrowding has become an issue

Refactored all existing commands to be more robust and less prone so permissions confusion (the bot couldn't tell the difference between a mod/admin and the actual tent host)

Changed the default settings for a tent to be more simple, bitrate will always be maxxed and the default user limit is 8. Imposed a default limit because overcrowding has become a problem particularly in tents where the host doesn't know how to use commands or is in too altered of a state

Removed hide command to reduce confusion surrounding how hiding interacts with everything else. Tents will still be hidden for people /tent banned