New metric: spamminess
Spamminess replaces "Distribution". This new metric is similar but takes in to account every user instead of just the user talking the most. Aims to heavily penalise situations where it is already busy and people are "spamming"

This combined with the "Density" metric helps balance the score. Density offsets a high spamminess score in a situation where spamminess may be high but the conversation only involves few users so the overall "busyness" isn't huge