Add discord sub support
This changes the way subscriber roles are handled.

Patreon (and now Discord) now gives a corresponding tier role when the user subscribes and tripbot is now responsible for giving out both the premium member role and the supporter role.

Updated the role removal logic to also handle removing of the supporter role when a subscriber is no longer subscribed.

Updated /donate with new info

Still needs some work as this is critical to get perfect on the first try.